change float to decimal value

+++ database/migrations/2018_01_12_093126_create_var_klimatologis_table.php
@@ -18,15 +18,15 @@ class CreateVarKlimatologisTable extends Migration
             $table->char('noticenumber_id',16)->unique()->index();
             $table->foreign('noticenumber_id')->references('noticenumber')->on('magma_vars')->onDelete('cascade');
             $table->char('cuaca',50)->nullable();
-            $table->float('curah_hujan',5,2)->default(0.0);
+            $table->decimal('curah_hujan',3,2)->default(0.0);
             $table->char('kecangin',32)->nullable();
             $table->char('arahangin',100)->nullable();
-            $table->float('suhumin',4,2)->default(0.0);
-            $table->float('suhumax',4,2)->default(0.0);
-            $table->float('lembabmin',4,2)->default(0.0);
-            $table->float('lembabmax',4,2)->default(0.0);
-            $table->float('tekmin',5,2)->default(0.0);
-            $table->float('tekmax',5,2)->default(0.0);
+            $table->decimal('suhumin',3,2)->default(0.0);
+            $table->decimal('suhumax',3,2)->default(0.0);
+            $table->decimal('lembabmin',3,2)->default(0.0);
+            $table->decimal('lembabmax',3,2)->default(0.0);
+            $table->decimal('tekmin',4,2)->default(0.0);
+            $table->decimal('tekmax',4,2)->default(0.0);
             $table->timestamps();
         });
     }
